调整coord Operator和command的接口
git-svn-id: http://192.168.20.11/sequoiadb/trunk@17915 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/cat/catDCLogMgr.cpp
@@ -331,7 +331,7 @@ namespace engine
          }
 
          // add objs to mb block
-         if( mb->idleSize() < buffObj.size() )
+         if( mb->idleSize() < (UINT32)buffObj.size() )
          {
             rc = mb->extend( buffObj.size() - mb->idleSize() ) ;
             PD_RC_CHECK( rc, PDERROR, "Failed to extend mb, rc: %d", rc ) ;
